+++ scr/Classes/Enemy.hpp
@@ -8,7 +8,7 @@
 
 class Enemy:public Entity {
 public:
-    Enemy(int enemyType,float life,nTPoint spawn,nTPoint size,vector<vector<GLuint> > animations,bool isHuman);
+    Enemy(int enemyType,double life,nTPoint spawn,nTPoint size,vector<vector<GLuint> > animations,bool isHuman);
     Enemy(const Enemy& orig);
     virtual ~Enemy();
 
@@ -21,12 +21,12 @@ public:
     static const int bossLife;
     static const nTPoint defaultSize;
     static const nTPoint bossSize;
-    static const float imunityTime;
+    static const double imunityTime;
     static vector<vector<string> >enemyAnim;
     static vector<vector<int> >enemyAnimSize;
 private:
     void stateControl();
-    void makeInvencible(float time);
+    void makeInvencible(double time);
 
     string nickname;
     static vector<vector<string> >nicks;
